fixed finallly!!!

+++ micro-service-panier/src/main/java/com/example/micro_service_panier/SAGA/SagaMessage.java
@@ -37,7 +37,7 @@ public class SagaMessage implements Serializable {
         this.dispo = dispo;
     }
 
-    public int getPanierId() {
+    public long getPanierId() {
         return panierId;
     }
